requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
Click==7.0
Flask==1.1.1
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.7.0
PyJWT==1.7.1
pylint==2.4.4
python-dotenv==0.11.0
pytz==2019.3
regex==2020.2.20
six==1.14.0
SQLAlchemy==1.3.13
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
